+++ examples/react/react-app/src/components/historical/historical.get-ticks.tsx
@@ -28,7 +28,7 @@ export function HistoricalGetTicks({
 
   const logErr = (...args: unknown[]) => setErrorLog(logs => {
     console.log(...args);
-    return [...logs, ...args];
+    return [...logs, ...args.map((arg: any) => arg.message || arg)];
   });
   const log = (...args: unknown[]) => setResultLog(logs => {
     console.log(...args);