+++ examples/react/react-app/src/components/stream-quotes/stream-quotes.tsx
@@ -33,7 +33,7 @@ export function StreamQuotes({
 
   const logErr = (...args: unknown[]) => setErrorLog(logs => {
     console.log(...args);
-    return [...logs, ...args];
+    return [...logs, ...args.map((arg: any) => arg.message || arg)];
   });
   const log = (...args: unknown[]) => setResultLog(logs => {
     console.log(...args);