+++ examples/vue/vue-app/src/components/copy-factory/copy-factory.telegram.vue
@@ -33,7 +33,7 @@ const setChatId = (newChatId: string) => { chatId.value = newChatId; };
 
 const logErr = (...args: unknown[]) => {
   console.log(...args);
-  errorLog.value = [...errorLog.value, ...args];
+  errorLog.value = [...errorLog.value, ...args.map((arg: any) => arg.message || arg)];
 };
 const log = (...args: unknown[]) => {
   console.log(...args);