+++ examples/vue/vue-app/src/components/risk-management/risk-management.equlity-balance.vue
@@ -23,7 +23,7 @@ const token = ref(props.defaultToken);
 
 const logErr = (...args: unknown[]) => {
   console.log(...args);
-  errorLog.value = [...errorLog.value, ...args];
+  errorLog.value = [...errorLog.value, ...args.map((arg: any) => arg.message || arg)];
 };
 const log = (...args: unknown[]) => {
   console.log(...args);